camera finally working

+++ app/src/main/java/net/cloudhacking/androidgame2/engine/utils/PointF.java
@@ -26,7 +26,7 @@ public class PointF {
     }
 
 
-    public void set(int x, int y) {
+    public void set(float x, float y) {
         this.x = x;
         this.y = y;
     }
